refactor(changelog): restructure package from src/ to changelog_generator/
Replace generic src/ directory with proper Python package structure:
- src/ → changelog_generator/ (Python package name)
- Improved readability and import clarity
- Enables future pip install compatibility
- Aligns with Python packaging conventions

Updates:
- cli.py: from src.agent → from changelog_generator.agent
- All tests: from src.* → from changelog_generator.*
- Patch decorators: @patch('src.agent') → @patch('changelog_generator.agent')
- Relative imports: unchanged (remain within package)

All 53 tests passing with new structure.
