+++ tools/changelog_generator/tests/test_models.py
@@ -8,7 +8,7 @@ import pytest
 # Add src directory to path
 sys.path.insert(0, str(Path(__file__).parent.parent))
 
-from src.models import Commit, ChangelogContent, ChangelogSection
+from changelog_generator.models import Commit, ChangelogContent, ChangelogSection
 
 
 class TestCommit:
